Strip out dash when relating SSH host and an app.
Application names have no dashes in their corresponding directory names,
so executing a drush command over SSH for an app where the SSH config
entry has a dash would fail, because the corresponding app dir does not
exist.

For instance:

    $ uli my-app.prod

would fail since we tried to `cd` to `~/apps/my-app/current` while in
fact we should cd to `~/apps/myapp/current`.

This is now fixed.

Thanks to @dubgeiser for the inspiration.