Update the ssh_drush command for Drush 12
The drush-launcher project has been deprecated for a while, and stopped
working since Drush 12. To upgrade to Drupal 10.2, we need to upgrade
drush to v12. This means the ssh dursh aliases would've stopped working
if we didn't update them to use the drush executable from the vendor
map. This also means that with this change, these aliases will no longer
work for the Drupal 7 sites, but these are a minority and will be
removed or upgraded soon.